Add files via upload
Buying new cosmetic products is difficult. It can even be scary for those who have sensitive skin andare prone to skin trouble. The information needed to alleviate this problem is on the back of eachproduct, but it's tought to interpret those ingredient lists unless you have a background in chemistry.Instead of buying and hoping for the best, we can use data science to help us predict which productsmay be good fits for us. In this Project, you are going to create a content-based recommendationsystem where the 'content' will be the chemical components of cosmetics. Specifically, you willprocess ingredient lists for 1472 cosmetics on Sephora via word embedding, then visualizeingredient similarity using a machine learning method called t-SNE and an interactive visualizationlibrary called Bokeh